ai/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

flask==3.0.3
flask-cors==4.0.0
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1
gunicorn==21.2.0
pyarrow==15.0.2
fastparquet==2024.2.0

# --- BAGIAN KRUSIAL (Downgrade agar kompatibel) ---
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
sentence-transformers==2.7.0
transformers==4.34.0